+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - The blob spell will now auto-undo after 30 seconds.
  - Some fixes to the Essentials integration.
  - Add /spellp command, for remotely casting a spell as another player.
+ - Add y_offset and y_value parameters to the FlingSpell.
 
 ## 2.8.0